Improve day5 solution
In terms of a coding challenge there's quite a lot of boilerplate for
conceputalising data, but it helps me so it stays.

The solution is now more performant. The previous iteration would check
the entire grid at the end for values coordinates that contain a value
>= 2 which would take N loops where N is the amount of points on the
grid. This solution stores a hashmap of coordinates that have a value >=
2, meaning we can find the score of the grid by just checking the length
of the hashmap.